requirements/dev.txt
-r base.txt

pytest==2.9.1
pytest-django==2.8.0
pytest-mock==0.11.0
